backend/requirements.txt
Django
psycopg2-binary
gunicorn
mysqlclient
google-cloud
google-cloud-vision
pymysql
django-apscheduler
apscheduler